+++ web/wiki/src/assets/icons/index.ts
@@ -1,5 +1,7 @@
 import angular from './angular.svg'
 import arrowDown from './arrow-down.svg'
+import bbdd from './bbdd.svg'
+import bgHomeCardCorner from './bg-home-card-corner.svg'
 import chevronDown from './chevron-down.svg'
 import dataScience from './data-science.svg'
 import deleteIcon from './delete-icon.svg'
@@ -8,6 +10,7 @@ import itLogo from './it-logo.svg'
 import java from './java.svg'
 import javascript from './javascript.svg'
 import lockDynamic from './lock-dynamic-color.svg'
+import logout from './logout.svg'
 import newFolder from './new-folder-dynamic-color.svg'
 import node from './nodejs-logo.svg'
 import php from './php.svg'
@@ -15,13 +18,13 @@ import profileAvatar from './profile-avatar.svg'
 import puzzleDynamic from './puzzle-dynamic-color.svg'
 import react from './react.svg'
 import thumbUp from './thumb-up-dynamic-color.svg'
-import logout from './logout.svg'
 import user from './user.svg'
-import bbdd from './bbdd.svg'
 
 const icons = {
   angular,
   arrowDown,
+  bbdd,
+  bgHomeCardCorner,
   chevronDown,
   dataScience,
   deleteIcon,
@@ -30,6 +33,7 @@ const icons = {
   java,
   javascript,
   lockDynamic,
+  logout,
   newFolder,
   node,
   php,
@@ -37,9 +41,7 @@ const icons = {
   puzzleDynamic,
   react,
   thumbUp,
-  logout,
   user,
-  bbdd,
 }
 
 export default icons